+++ src/django/api/tests.py
@@ -540,8 +540,9 @@ class GeocodingUtilsTest(TestCase):
         )
 
     def test_geocoded_address_data_is_formatted_correctly(self):
+        data = parsed_city_hall_data['full_response']
         formatted_data = format_geocoded_address_data(
-            parsed_city_hall_data['full_response'])
+            data, data['results'][0])
         self.assertEqual(formatted_data, parsed_city_hall_data)
 
 
@@ -573,6 +574,142 @@ geocoding_data = {'results': [{'address_components': [
     'place_id': 'ChIJ8cV_ZH_IxokRA_ETpdB5R3Y',
     'types': ['premise']}],
     'status': 'OK'}
+
+geocoding_data_no_country = {
+   "results": [
+      {
+         "address_components": [
+            {
+               "long_name": "4WHM+QCX",
+               "short_name": "4WHM+QCX",
+               "types": ["plus_code"]
+            },
+            {
+               "long_name": "Famagusta",
+               "short_name": "Famagusta",
+               "types": ["locality", "political"]
+            },
+            {
+               "long_name": "99450",
+               "short_name": "99450",
+               "types": ["postal_code"]
+            }
+         ],
+         "formatted_address": "4WHM+QCX, Famagusta 99450",
+         "geometry": {
+            "location": {"lat": 35.1294866, "lng": 33.9336133},
+            "location_type": "GEOMETRIC_CENTER",
+            "viewport": {
+               "northeast": {
+                  "lat": 35.1308355802915,
+                  "lng": 33.9349622802915
+               },
+               "southwest": {
+                  "lat": 35.1281376197085,
+                  "lng": 33.9322643197085
+               }
+            }
+         },
+         "partial_match": True,
+         "place_id": "ChIJDzTqezLI3xQRW5kjGEGLRzA",
+         "types": ["establishment", "point_of_interest"]
+      }
+   ],
+   "status": "OK"
+}
+
+geocoding_data_second_country = {
+   "results": [
+      {
+         "address_components": [
+            {
+               "long_name": "Noor Bagh",
+               "short_name": "Noor Bagh",
+               "types": ["political", "sublocality", "sublocality_level_1"]
+            },
+            {
+               "long_name": "Srinagar",
+               "short_name": "Srinagar",
+               "types": ["locality", "political"]
+            },
+            {
+               "long_name": "190009",
+               "short_name": "190009",
+               "types": ["postal_code"]
+            }
+         ],
+         "formatted_address": "Noor Bagh, Srinagar 190009",
+         "geometry": {
+            "bounds": {
+               "northeast": {"lat": 34.07490560000001, "lng": 74.8043599},
+               "southwest": {"lat": 34.0734423, "lng": 74.8023449}
+            },
+            "location": {"lat": 34.0742387, "lng": 74.8035549},
+            "location_type": "APPROXIMATE",
+            "viewport": {
+               "northeast": {
+                  "lat": 34.07552293029151,
+                  "lng": 74.8047013802915
+               },
+               "southwest": {
+                  "lat": 34.07282496970851,
+                  "lng": 74.80200341970848
+               }
+            }
+         },
+         "partial_match": True,
+         "place_id": "ChIJmQlkN-2P4TgR6u3glWeOMTE",
+         "types": ["political", "sublocality", "sublocality_level_1"]
+      },
+      {
+         "address_components": [
+            {
+               "long_name": "Kaliakair",
+               "short_name": "Kaliakair",
+               "types": ["locality", "political"]
+            },
+            {
+               "long_name": "Gazipur District",
+               "short_name": "Gazipur District",
+               "types": ["administrative_area_level_2", "political"]
+            },
+            {
+               "long_name": "Dhaka Division",
+               "short_name": "Dhaka Division",
+               "types": ["administrative_area_level_1", "political"]
+            },
+            {
+               "long_name": "Bangladesh",
+               "short_name": "BD",
+               "types": ["country", "political"]
+            }
+         ],
+         "formatted_address": "Kaliakair, Bangladesh",
+         "geometry": {
+            "bounds": {
+               "northeast": {"lat": 24.0840819, "lng": 90.2365494},
+               "southwest": {"lat": 24.0535966, "lng": 90.2030754}
+            },
+            "location": {"lat": 24.0694528, "lng": 90.2221213},
+            "location_type": "APPROXIMATE",
+            "viewport": {
+               "northeast": {
+                  "lat": 24.0840819,
+                  "lng": 90.2365494
+               },
+               "southwest": {
+                  "lat": 24.0535966,
+                  "lng": 90.2030754
+               }
+            }
+         },
+         "partial_match":  True,
+         "place_id": "ChIJTfs3ierjVTcRysKYnbOKmZM",
+         "types": ["locality", "political"]
+      }
+   ],
+   "status": "OK"
+}
 
 
 class GeocodingTest(TestCase):
@@ -606,8 +743,38 @@ class GeocodingTest(TestCase):
 
         self.assertEqual(
             cm.exception.args,
-            ("Geocoding results did not match provided country code.",)
+            ("Geocoding results of US did not match " +
+             "provided country code of IN.",)
         )
+
+    @patch('api.geocoding.requests.get')
+    def test_accepts_inexact_address(self, mock_get):
+        mock_get.return_value = Mock(ok=True, status_code=200)
+        mock_get.return_value.json.return_value = geocoding_data_no_country
+
+        expected_result = geocoding_data_no_country['results'][0]
+        expected_point = expected_result["geometry"]["location"]
+        expected_address = expected_result["formatted_address"]
+
+        results = geocode_address('PortİSBİ Serbest Bölge Office:4, ' +
+                                  'Gazimağusa, North Cyprus, 99450',
+                                  'TR')
+        self.assertEqual(expected_point, results["geocoded_point"])
+        self.assertEqual(expected_address, results['geocoded_address'])
+
+    @patch('api.geocoding.requests.get')
+    def test_accepts_alternate_address(self, mock_get):
+        mock_get.return_value = Mock(ok=True, status_code=200)
+        mock_get.return_value.json.return_value = geocoding_data_second_country
+
+        expected_result = geocoding_data_second_country['results'][1]
+        expected_point = expected_result["geometry"]["location"]
+        expected_address = expected_result["formatted_address"]
+
+        results = geocode_address('Noorbagh, Kaliakoir Gazipur Dhaka 1704',
+                                  'BD')
+        self.assertEqual(expected_point, results["geocoded_point"])
+        self.assertEqual(expected_address, results['geocoded_address'])
 
 
 class FacilityAndProcessingTypeTest(TestCase):
@@ -672,6 +839,101 @@ class FacilityAndProcessingTypeTest(TestCase):
         self.assertEqual(output, expected_output)
 
 
+listitem_geocode_data = {
+   "results": [
+      {
+         "address_components": [
+            {
+               "long_name": "Linjiacunzhen",
+               "short_name": "Linjiacunzhen",
+               "types": ["political", "sublocality", "sublocality_level_2"]
+            },
+            {
+               "long_name": "Zhucheng",
+               "short_name": "Zhucheng",
+               "types": ["political", "sublocality", "sublocality_level_1"]
+            },
+            {
+               "long_name": "Weifang",
+               "short_name": "Weifang",
+               "types": ["locality", "political"]
+            },
+            {
+               "long_name": "Shandong",
+               "short_name": "Shandong",
+               "types": ["administrative_area_level_1", "political"]
+            },
+            {
+               "long_name": "China",
+               "short_name": "CN",
+               "types": ["country", "political"]
+            },
+            {
+               "long_name": "262232",
+               "short_name": "262232",
+               "types": ["postal_code"]
+            }
+         ],
+         "formatted_address": "Linjiacunzhen, Zhucheng, " +
+                              "Weifang, Shandong, China, 262232",
+         "geometry": {
+            "location": {"lat": 35.994813, "lng": 119.65418},
+            "location_type": "APPROXIMATE",
+            "viewport": {
+               "northeast": {"lat": 36.0038401, "lng": 119.6701874},
+               "southwest": {"lat": 35.9857849, "lng": 119.6381726}
+            }
+         },
+         "partial_match": True,
+         "place_id": "ChIJV5SPiTEkvjUR7ErhmnSRTR8",
+         "types": ["political", "sublocality", "sublocality_level_2"]
+      },
+      {
+         "address_components": [
+            {
+               "long_name": "396210",
+               "short_name": "396210",
+               "types": ["postal_code"]
+            },
+            {
+               "long_name": "Daman",
+               "short_name": "Daman",
+               "types": ["administrative_area_level_2", "political"]
+            },
+            {
+               "long_name": "Dadra and Nagar Haveli and Daman and Diu",
+               "short_name": "DH",
+               "types": ["administrative_area_level_1", "political"]
+            },
+            {
+               "long_name": "India",
+               "short_name": "IN",
+               "types": ["country", "political"]
+            }
+         ],
+         "formatted_address":  "Dadra and Nagar Haveli and Daman and " +
+                               "Diu 396210, India",
+         "geometry": {
+            "bounds": {
+               "northeast": {"lat": 20.4696847, "lng": 72.8751228},
+               "southwest": {"lat": 20.4051972, "lng": 72.82805549999999}
+            },
+            "location": {"lat": 20.4346424, "lng": 72.8456399},
+            "location_type": "APPROXIMATE",
+            "viewport": {
+               "northeast": {"lat": 20.4696847, "lng": 72.8751228},
+               "southwest": {"lat": 20.4051972, "lng": 72.82805549999999}
+            }
+         },
+         "partial_match": True,
+         "place_id": "ChIJ8d4EeIra4DsR3xkUSZh_-ng",
+         "types": ["postal_code"]
+      }
+   ],
+   "status": "OK"
+}
+
+
 class FacilityListItemGeocodingTest(ProcessingTestCase):
     def test_invalid_argument_raises_error(self):
         with self.assertRaises(ValueError) as cm:
@@ -699,7 +961,10 @@ class FacilityListItemGeocodingTest(ProcessingTestCase):
             ('Items to be geocoded must be in the PARSED status',),
         )
 
-    def test_incorrect_country_code_has_error_status(self):
+    @patch('api.geocoding.requests.get')
+    def test_nested_correct_country_code_succeeds(self, mock_get):
+        mock_get.return_value = Mock(ok=True, status_code=200)
+        mock_get.return_value.json.return_value = listitem_geocode_data
         facility_list = FacilityList.objects.create(
             header='address,country,name')
         source = Source.objects.create(
@@ -708,6 +973,30 @@ class FacilityListItemGeocodingTest(ProcessingTestCase):
         item = FacilityListItem(
             raw_data='"Linjiacun Town, Zhucheng City Weifang, Daman, ' +
                      'Daman, 396210",IN,Shirts!',
+            source=source
+        )
+        parse_facility_list_item(item)
+        geocode_facility_list_item(item)
+
+        expected_result = listitem_geocode_data['results'][1]
+        expected_address = expected_result['formatted_address']
+
+        self.assertEqual(item.status, FacilityListItem.GEOCODED)
+        self.assertEqual(item.geocoded_address, expected_address)
+        self.assertIsInstance(item.geocoded_point, Point)
+
+    @patch('api.geocoding.requests.get')
+    def test_incorrect_country_code_has_error_status(self, mock_get):
+        mock_get.return_value = Mock(ok=True, status_code=200)
+        mock_get.return_value.json.return_value = listitem_geocode_data
+        facility_list = FacilityList.objects.create(
+            header='address,country,name')
+        source = Source.objects.create(
+            source_type=Source.LIST,
+            facility_list=facility_list)
+        item = FacilityListItem(
+            raw_data='"Linjiacun Town, Zhucheng City Weifang, Daman, ' +
+                     'Daman, 396210",BD,Shirts!',
             source=source
         )
         parse_facility_list_item(item)